# src/main/resources/openapi/auth-api-v1.yaml
openapi: 3.0.0
info:
  title: Nutrisoft Authentication API
  description: API for user authentication and authorization (patients, professionals)
  version: 1.0.0
  contact:
    name: Nutrisoft Support
    email: support@nutrisoft.com
  license:
    name: Apache 2.0

servers:
  - url: http://localhost:8080/api
    description: Development server

tags:
  - name: Authentication
    description: User authentication and registration operations (professionals only)

paths:
  /v1/auth/login:
    post:
      summary: Login user
      description: Authenticates a user and returns a JWT token
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful, JWT token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid credentials or inactive account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/auth/refresh-token:
    post:
      summary: Refresh access token
      description: |
        Renueva el access token usando el refresh token enviado como cookie HttpOnly.
        El refresh token debe estar presente en la cookie 'refreshToken'.
      operationId: refreshToken
      tags:
        - Authentication
      parameters:
        - name: refreshToken
          in: cookie
          description: HttpOnly cookie containing the refresh token JWT
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Nuevo access token generado correctamente. Incluye nueva cookie de refresh token.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: New refreshToken cookie (HttpOnly, Secure, SameSite=Strict)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Refresh token inválido, ausente, expirado o revocado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                no_token:
                  value:
                    error: Unauthorized
                    message: Refresh token not provided
                invalid_token:
                  value:
                    error: Unauthorized
                    message: Refresh token is invalid or has expired
                wrong_type:
                  value:
                    error: Unauthorized
                    message: Token type is not a refresh token
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/auth/logout:
    post:
      summary: Logout user
      description: |
        Revoca el refresh token actual y limpia la cookie.
        Esto cierra la sesión actual del usuario.
      operationId: logout
      tags:
        - Authentication
      parameters:
        - name: refreshToken
          in: cookie
          description: HttpOnly cookie containing the refresh token JWT
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Logout exitoso. Cookie de refresh token limpiada.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: refreshToken cookie cleared (expired)
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User password
          example: SecurePassword123!

    AuthResponse:
      type: object
      required:
        - token
        - type
        - email
        - role
      properties:
        token:
          type: string
          description: JWT Bearer token for authentication
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type:
          type: string
          default: Bearer
          description: Token type
          example: Bearer
        aggregateId:
          type: string
          format: uuid
          nullable: true
          description: Aggregate ID of the professional
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          description: User email
          example: user@example.com
        role:
          type: string
          enum:
            - PATIENT
            - PROFESSIONAL
            - ADMIN
          description: User role
          example: PROFESSIONAL

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type or title
          example: Authentication failed
        message:
          type: string
          description: Detailed error message
          example: Invalid email or password

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token

security:
  - bearerAuth: []